fix&test: 增加entry和transfer的测试，并修复transfer表的主键定义问题

+++ db/migration/000001_init_schema.up.sql
@@ -14,7 +14,7 @@ CREATE TABLE "entries" (
 );
 
 CREATE TABLE "transfers" (
-  "id" int PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "from_account_id" bigint NOT NULL,
   "to_account_id" bigint NOT NULL,
   "amount" bigint NOT NULL,
